P1 C2: Move discovery.rb → inventory/discovery.rb (move-only)
Module path: Legion::LLM::Discovery → Legion::LLM::Inventory::Discovery.
Old path lib/legion/llm/discovery.rb becomes a require_relative forwarder.
compat.rb const_missing alias Discovery → Inventory::Discovery for cross-gem callers.
In-gem refs updated: inventory.rb now uses Legion::LLM::Inventory::Discovery.
System.* calls use fully-qualified path until commit 4 moves system.rb.
